ISSUE-9: extract autodetect executable into separate module
